+++ VSCode/Elektrogrrosshandel/Elektrogrrosshandel/Hardware/500_GraphicsCard.cs
@@ -50,7 +50,7 @@ namespace Elektrogrosshandel.Hardware
         private static int ArticelGroupID = 500;
         private string ArticelGroupDescription = "This category includes discrete graphics cards with various memory sizes, clock rates and output configurations.";
 
-        private static List<int> ArticelIDs = new List<int>();
+        private static List<Int64> ArticelIDs = new List<Int64>();
 
         public GraphicsCard(string articelName, string articelManufacturer, string articelModel,
                               int articelYearOfProduction, int articelManufactrerID, string[] articelColors, int articelStock,
@@ -76,10 +76,10 @@ namespace Elektrogrosshandel.Hardware
 
         }
 
-        private static int CreateArticelID()
+        private static Int64 CreateArticelID()
         {
             string articelID;
-            int iD;
+            Int64 iD;
             Random random = new Random();
             do
             {
@@ -91,7 +91,7 @@ namespace Elektrogrosshandel.Hardware
             } while (true);
 
             articelID = ComputerHardware.ArticelParentGroupID + ArticelGroupID.ToString() + iD.ToString("D4");
-            iD = int.Parse(articelID);
+            iD = Int64.Parse(articelID);
             ArticelIDs.Add(iD);
 
             return iD;
