+++ VSCode/Elektrogrrosshandel/Elektrogrrosshandel/Hardware/600_Ram.cs
@@ -40,7 +40,7 @@ namespace Elektrogrosshandel.Hardware
         private static int ArticelGroupID = 600;
         private string ArticelGroupDescription = "Diese Kategorie enthält Arbeitsspeicher (RAM) für Computer in verschiedenen Kapazitäten, Frequenzen und Typen (z. B. DDR4, DDR5).";
 
-        private static List<int> ArticelIDs = new List<int>();
+        private static List<Int64> ArticelIDs = new List<Int64>();
 
         public Ram(string articelName, string articelManufacturer, string articelModel,
                     int articelYearOfProduction, int articelManufactrerID, string[] articelColors, int articelStock,
@@ -61,10 +61,10 @@ namespace Elektrogrosshandel.Hardware
             ComputerHardware.AddRAM(this);
         }
 
-        private static int CreateArticelID()
+        private static Int64 CreateArticelID()
         {
             string articelID;
-            int iD;
+            Int64 iD;
             Random random = new Random();
             do
             {
@@ -77,7 +77,7 @@ namespace Elektrogrosshandel.Hardware
             } while (true);
 
             articelID = ComputerHardware.ArticelParentGroupID + ArticelGroupID.ToString() + iD.ToString("D4");
-            iD = int.Parse(articelID);
+            iD = Int64.Parse(articelID);
             ArticelIDs.Add(iD);
 
             return iD;
