+++ VSCode/Elektrogrrosshandel/Elektrogrrosshandel/Hardware/700_StorageDevices.cs
@@ -76,7 +76,7 @@ namespace Elektrogrosshandel.Hardware
         private static int ArticelGroupID = 700;
         private string ArticelGroupDescription = "This category includes storage devices such as HDDs and SSDs with various interfaces and form factors.";
 
-        private static List<int> ArticelIDs = new List<int>();
+        private static List<Int64> ArticelIDs = new List<Int64>();
 
         public StorageDevice(string articelName, string articelManufacturer, string articelModel,
                                int articelYearOfProduction, int articelManufactrerID, string[] articelColors, int articelStock,
@@ -99,10 +99,10 @@ namespace Elektrogrosshandel.Hardware
             ComputerHardware.AddStorageDevice(this);
         }
 
-        private static int CreateArticelID()
+        private static Int64 CreateArticelID()
         {
             string articelID;
-            int iD;
+            Int64 iD;
             Random random = new Random();
             do
             {
@@ -114,7 +114,7 @@ namespace Elektrogrosshandel.Hardware
             } while (true);
 
             articelID = ComputerHardware.ArticelParentGroupID + ArticelGroupID.ToString() + iD.ToString("D4");
-            iD = int.Parse(articelID);
+            iD = Int64.Parse(articelID);
             ArticelIDs.Add(iD);
 
             return iD;
