+++ VSCode/Elektrogrrosshandel/Elektrogrrosshandel/Hardware/950_Software.cs
@@ -61,7 +61,7 @@ namespace Elektrogrosshandel.Hardware
         private static int ArticelGroupID = 950;
         private string ArticelGroupDescription = "Softwareprodukte, Lizenzen und digitale Güter.";
 
-        private static List<int> ArticelIDs = new List<int>();
+        private static List<Int64> ArticelIDs = new List<Int64>();
 
         public Software(string articelName, string articelManufacturer, string articelModel,
                         int articelYearOfProduction, int articelManufactrerID, string[] articelColors, int articelStock,
@@ -77,10 +77,10 @@ namespace Elektrogrosshandel.Hardware
             ComputerHardware.AddSoftware(this);
         }
 
-        private static int CreateArticelID()
+        private static Int64 CreateArticelID()
         {
             string articelID;
-            int iD;
+            Int64 iD;
             Random random = new Random();
             do
             {
@@ -92,7 +92,7 @@ namespace Elektrogrosshandel.Hardware
             } while (true);
 
             articelID = ComputerHardware.ArticelParentGroupID + ArticelGroupID.ToString() + iD.ToString("D4");
-            iD = int.Parse(articelID);
+            iD = Int64.Parse(articelID);
             ArticelIDs.Add(iD);
 
             return iD;
